requirements.txt
llama-cpp-python
tqdm
IPython
unidecode
requests
openai